fix: sometimes pages cannot be created nor updated (#5)
* fix(editablefield): add resetTrigger to reset the Editor on discard

Because of EditableField content do not update when the editing content changes the useEffect of the
Editor is not triggered on discard. By adding a resetTrigger this issue has been fixed.

* fix(editor): cursor no longer jumps to left side on editing

When editing a StreamField the cursor now no longer jumps to the left side which made editing
impossible.

* fix(reducers/cms): registerPage now correctly take care of the parent page

* fix(reducers/cms): registerPage calculates the correct details diff now

registerPage now calculates the correct difference between editingLayer, pageDetails and current
data.