feat(datalayer): add encryption for WDL files
Changes the working datalayer structure by implemented a `crypt` object.
Crypt consits of a `clear` and `cipher` section, where `cipher` includes
an encrypted reflaction of the `clear` structure.
When you log in, `cipher` is decrypted and overrides `clear`.

This change is introduced for data protection reasons in order not to
make all files public (exceptions possible).