perf: reduce layer diff. calc. costs (#13)
The module `react-diff-viewer` cannot handle large inputs, so we need to
change the way we calculate the diffs. This is done by the google diff
module.
The way of displaying the diff is not optimal yet, so we need to change
it in the future by improving the `diffPrettyHtml` function.