fix(fields): add optional pageId to fields
This feature should already be included into v2.0.0 but went missing.
This change adds a optional pageId property to all fields thatare based
on the template id of the jaen context.

This change is required to make the fields more flexible by allowing
to use the same field in different contexts.