feat: add page duplication functionality
This commit adds the ability to duplicate a page and its subpages. The user can now click on the "Duplicate page" button in the danger zone actions of a page, enter a new slug for the duplicated page, and confirm the duplication. The duplicated page will have a slug based on the original page's slug with "-copy" appended to it. The title of the duplicated page will also have "(copy)" appended to it. If there are any child pages associated with the original page, they will also be duplicated recursively.

The code changes include:
- Importing `FaClone` icon from `@react-icons/all-files/fa/FaClone`
- Adding an `IconButton` component import
- Adding a new action object for duplicating a page in `dangerZoneActions`
- Implementing logic for duplicating a page in `clonePage` function in `CMSManagementContext`
- Handling success and error notifications using `toast`

Note: This commit does not include specific file names or paths from the codebase.