feat(cms): add default date value for blog posts
This commit adds a default date value for blog posts if the `date` field is not provided. The `PageContentForm` component now watches the `blogPost.date` field and sets it to the current date and time if it is empty. This ensures that all blog posts have a valid date when submitted.